ext4 crypto: sync up the mainline 4.1-rc1 merge
There were some last minute changes that weren't reflected in the ext4
crypto patches that we were syncing with flounder.  They were mostly
whitespace changes, plus an error handling bugfix if there was a
normal (non-crypto-related) bugs when adding a directory entry to an
inode while creating a file.

Change-Id: I01e1f8ee07aef2f826a27efcbfa85a825000f2bc
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Theodore Ts'o <tytso@google.com>

Conflicts:
	fs/ext4/crypto.c
	fs/ext4/crypto_fname.c
	fs/ext4/ext4.h
	fs/ext4/ioctl.c
	fs/ext4/namei.c